ci: restore stylelint, ignore micromatch CVE (#639)
This reverts commit 19ab2eeb9c5149acf2dd1844b29710e7e4e4685f.

This adds an `auditConfig` to ignore CVE-2024-4067, which is for a vulnerability in `micromatch`, a nested dependency of `stylelint`. This does not run in production. It does run in CI, but as far as I can tell `stylelint` does not hit `micromatch`'s vulnerable code path (it doesn't call `micromatch.parse` or `micromatch.braces`).